refactor: minimize changes to core crates
Further simplification to reduce impact on existing crates:

## Moved BuiltinsExt to starpls_stubs
- Remove BuiltinsExt trait from starpls_bazel (zero changes to core crate)
- Move merge functionality to starpls_stubs where it's actually used
- Keep starpls_bazel completely untouched

## Removed unnecessary dialect.rs
- Delete unused dialect.rs file from starpls_bazel
- Remove leftover code from complex extension system
- Keep starpls_bazel minimal and focused

## Cleaned up HIR trait
- Remove outdated comments about 'file-specific prelude methods'
- Keep trait definition clean and minimal
- Only essential changes remain

## Benefits
- Zero changes to starpls_bazel (core crate untouched)
- Minimal changes to starpls_hir (just trait cleanup)
- starpls_common only has dependency changes
- All functionality preserved with maximum isolation
