+++ crates/starpls/src/commands/check.rs
@@ -14,8 +14,7 @@ use anyhow::bail;
 use clap::Args;
 use starpls_bazel::client::BazelCLI;
 use starpls_bazel::client::BazelInfo;
-use starpls_bazel::Builtins;
-use starpls_bazel::BuiltinsExt;
+use starpls_stubs::BuiltinsExt;
 use starpls_common::Diagnostic;
 use starpls_common::Dialect;
 use starpls_common::FileId;
