+++ crates/starpls_hir/src/lib.rs
@@ -110,14 +110,10 @@ const TARGET_DOC: &str = "The BUILD target for a dependency. Appears in the fiel
 pub trait Db: salsa::DbWithJar<Jar> + starpls_common::Db {
     fn gcx(&self) -> &GlobalContext;
 
-    // Legacy dialect methods (for backward compatibility)
     fn set_builtin_defs(&mut self, dialect: Dialect, builtins: Builtins, rules: Builtins);
     fn get_builtin_defs(&self, dialect: &Dialect) -> BuiltinDefs;
-
     fn set_bazel_prelude_file(&mut self, file_id: FileId);
     fn get_bazel_prelude_file(&self) -> Option<FileId>;
-
-    // File-specific prelude methods for extensions
     fn set_all_workspace_targets(&mut self, targets: Vec<String>);
     fn get_all_workspace_targets(&self) -> Arc<Vec<String>>;
 }
